requirements.txt
certifi==2023.5.7 ; python_version >= "3.9"
charset-normalizer==3.1.0 ; python_version >= "3.9"
click==8.1.3 ; python_version >= "3.9"
colorama==0.4.6 ; python_version >= "3.9" and platform_system == "Windows"
idna==3.4 ; python_version >= "3.9"
python-dotenv==1.0.0 ; python_version >= "3.9"
requests==2.31.0 ; python_version >= "3.9"
urllib3==2.0.2 ; python_version >= "3.9"
